[add] 알림메시지 생성 훅 추가
