+++ libs/shared-types/src/index.ts
@@ -22,6 +22,7 @@ export * from './lib/dto/adminSettlementInfo.dto';
 export * from './lib/dto/broadcaster.dto';
 export * from './lib/dto/broadcasterAddress.dto';
 export * from './lib/dto/broadcasterContact.dto';
+export * from './lib/dto/broadcasterContractionAgreement.dto';
 export * from './lib/dto/broadcasterPurchase.dto';
 export * from './lib/dto/broadcasterSettlementInfo.dto';
 export * from './lib/dto/broadcasterSettlementInfoConfirmation.dto';
@@ -32,6 +33,7 @@ export * from './lib/dto/changeGoodsView.dto';
 export * from './lib/dto/changeNickname.dto';
 export * from './lib/dto/changeReturnStatus.dto';
 export * from './lib/dto/changeSellCommission.dto';
+export * from './lib/dto/createAdminNotification.dto';
 export * from './lib/dto/createBroadcasterChannel.dto';
 export * from './lib/dto/deleteGoods.dto';
 export * from './lib/dto/emailCodeVerification.dto';
@@ -88,4 +90,3 @@ export * from './lib/res-types/orderCancelRequest.res';
 export * from './lib/res-types/overlayControllerMain.res';
 export * from './lib/res-types/socialAccounts.res';
 export * from './lib/res-types/userProfile.res';
-export * from './lib/dto/broadcasterContractionAgreement.dto';
